Move LogFace type to settings
